services/api_v1/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
httpx==0.25.2
pydantic==2.5.0
